Version 0.28
Alpha test

Notice:
In the files "telegram.py" and "vkontakte.py". and also partially in "server.py" there have been changes related to the bugfix.
Bots don't hang now and work as they should.
Soon the bugfix will affect and "server.py", at least when the work on this file is finished.